🎨 Design: 프로젝트 수정 룰 추가 및 파일 구조 변경
