+++ src/components/admin/project/ProjectEditRule.jsx
@@ -1,0 +1,10 @@
+export default function ProjectEditRule() {
+  return (
+    <div className="flex flex-col w-92 h-36 border bg-[#F9F9F9] justify-center ml-auto pl-5 text-[0.9rem] font-semibold">
+      <p>1. 수정하고 싶은 필드 눌러 내용 작성 </p>
+      <p>2. 이미지 여러 개 동시 첨부 가능</p>
+      <p>3. 수정하고 싶은 트랙 수정 가능 </p>
+      <p>4. 이름 입력 후 엔터 입력 시 다음 이름 입력 가능</p>
+    </div>
+  );
+}
